requirements_without_whisper.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
colorama==0.4.6
distro==1.9.0
ffmpeg-python==0.2.0
filelock==3.16.1
fsspec==2024.10.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.7.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
openai==1.54.4
pi==0.1.2
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
rsa==4.9
setuptools==75.5.0
sniffio==1.3.1
sympy==1.13.1
Telethon==1.37.0
torch==2.5.1
tqdm==4.67.0
typing_extensions==4.12.2